+++ Projeto_PDSI_II/src/Cadastros/Cadastro_Alimentos_Vegetais.java
@@ -137,8 +137,8 @@ public class Cadastro_Alimentos_Vegetais extends JFrame implements ActionListene
                 
             stmt.setString(1,dados_V.getTipo());
             stmt.setString(2,dados_V.getFornecedor());
-            stmt.setFloat(3,dados_V.getQuantidade());
-            stmt.setFloat(4,dados_V.getPreco());
+            stmt.setDouble(3,dados_V.getQuantidade());
+            stmt.setDouble(4,dados_V.getPreco());
          
         } catch (SQLException ex) {
